reg/login logic refactoring
