+++ package.json
@@ -16,7 +16,6 @@
     "@coreui/coreui-plugin-chartjs-custom-tooltips": "^1.3.1",
     "@coreui/icons": "^0.3.0",
     "@coreui/react": "^2.5.4",
-    "axios": "^0.19.0",
     "bootstrap": "^4.3.1",
     "chart.js": "^2.9.3",
     "classnames": "^2.2.6",
@@ -39,7 +38,6 @@
     "react-router-dom": "^5.1.2",
     "react-table": "^6.10.3",
     "react-test-renderer": "^16.12.0",
-    "react-yandex-maps": "^4.2.0",
     "reactstrap": "^8.1.1",
     "simple-line-icons": "^2.4.1",
     "socket.io": "^2.3.0",
